+++ packages/terra-clinical-label-value-view/CHANGELOG.md
@@ -3,6 +3,7 @@ ChangeLog
 
 Unreleased
 ----------
+* Updated value font-size and the site example page
 
 2.1.0 - (March 6, 2018)
 ----------